CJAL-05/10/2018. Se finaliza con las vistas del modulo Caja general, asi como pequeñas correciones de los modulos anteriores.

+++ 1.Layers/1.1Presentation/1.1.1Web/MVC.Presentacion/Controllers/PrecioVentaOtroController.cs
@@ -69,7 +69,7 @@ namespace MVC.Presentacion.Controllers
             }
 
         }
-        public ActionResult EditarPrecioVentaOtro(int id)
+        public ActionResult EditarPrecioVentaOtro(short id)
         {
             if (Session["StringToken"] == null) return RedirectToAction("Index", "Home", AutenticacionServicio.InitIndex(new Models.Seguridad.LoginModel()));
             string _tkn = Session["StringToken"].ToString();
